requirements.txt
einops
tqdm
ruamel.yaml
wandb
timm
scikit-learn
webdataset
librosa>=0.10.0
soundfile
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers
torchlibrosa
pyloudnorm
pystoi
torcheval
torchmetrics
